* move Label class into a separate file, so do the tests for Label class * rearrange the imports in test_img_feature.py
